backend/pyproject.toml
[tool.poetry]
name = "umanager"
version = "0.1.0"
description = "Backend for uManager"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.20"
pydantic-settings = "^2.0.3"
email-validator = "^2.0.0.post2"
psycopg2-binary = "^2.9.7"
alembic = "^1.12.0"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
tenacity = "^8.2.3"
pillow = "^10.0.1"
authlib = "^1.2.1"
httpx = "^0.25.0"
itsdangerous = "^2.1.2"
jinja2 = "^3.1.2"
fastapi-mail = "^1.4.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
